pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mfai_db_repos"
version = "0.1.0"
authors = [
    {name = "MODFLOW-AI Team"}
]
description = "A repository indexing system with vector embeddings"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "SQLAlchemy",
    "psycopg2-binary",
    "pgvector",  # PostgreSQL vector extension for SQLAlchemy
    "GitPython",
    "click",
    "pydantic",
    "openai",
    "google-genai",
    "rich",
    "asyncio",
    "asyncpg",
    "aiohttp",
    "python-dotenv",
    "chardet",
    "python-magic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "black",
    "isort",
    "ruff",
]

[tool.setuptools]
packages = ["mfai_db_repos"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true